# feat: Add smart TypeScript validation with context-aware performance optimization