--- a/package.json
+++ b/package.json
@@ -36,6 +36,11 @@
     "typecheck": "tsc --project tsconfig.build.json --noEmit",
     "typecheck:fast": "tsc --project tsconfig.build.json --noEmit --skipLibCheck",
     "typecheck:quick": "node scripts/quick-typecheck.js",
+    "check": "node scripts/smart-check.js",
+    "check:dev": "npm run typecheck:quick && npm run lint",
+    "check:pr": "npm run typecheck:quick && npm run lint && node run-basic-tests.js",
+    "check:full": "npm run typecheck && npm run lint && node run-basic-tests.js && npm run build",
+    "show-checks": "node scripts/show-checks.js",
     "prepare": "husky"
   },
   "dependencies": {