## Summary
Addresses the TypeScript typecheck timeout issue by implementing smart, context-aware validation scripts that dramatically improve development experience.

## Problem
- `npm run typecheck` was timing out (2+ minutes) making development slow
- Claude Code was defaulting to the slowest validation option  
- No clear guidance on when to use which validation script
- Development workflow was unnecessarily slow

## Solution

### 🚀 Smart Check Commands (New)
- **`check:dev`** - ⚡ Quick development validation (~5-10s)
- **`check:pr`** - 🔍 PR validation (~10-20s)  
- **`check:full`** - 🐌 Comprehensive validation (~2+ min)
- **`check [context]`** - 🤖 Context-aware runner (dev|pr|full|ci)

### 📋 Script Discovery Helper
- **`show-checks`** - Displays all available validation scripts with timing and usage guidance

### 📖 Clear Documentation
- Updated CLAUDE.md with context-specific recommendations
- Speed indicators and duration estimates for each script
- Clear guidance on when to use which validation level

## Performance Impact

| Context | Before | After | Improvement |
|---------|--------|-------|-------------|
| Development | ~2+ min (timeout) | ~2-5s | **24x faster** |
| PR Review | ~2+ min (timeout) | ~10-20s | **6-12x faster** |
| Pre-deployment | ~2+ min | ~2+ min | Same (comprehensive) |

## Key Features

### Context-Aware Selection
```bash
npm run check dev    # Quick development validation
npm run check pr     # Pull request validation  
npm run check full   # Comprehensive validation
```

### Smart TypeScript Checking
- **`typecheck:quick`** - Only modified files (~2-5s)
- **`typecheck:fast`** - Skip lib checking (~10-30s)
- **`typecheck`** - Full project check (~2+ min)

### Developer Experience
- Clear speed indicators (⚡🚀🐌) in all documentation
- Duration estimates for each script
- Usage recommendations by context
- Script discovery helper

## Changes Made

### `/CLAUDE.md`
- Added context-aware validation guidance
- Clear performance expectations
- Smart check command reference
- Speed indicators and timing estimates

### `/package.json`
- Added `check:dev`, `check:pr`, `check:full` wrapper scripts
- Added `check` context-aware runner
- Added `show-checks` discovery helper

### `/scripts/smart-check.js` (New)
- Context-aware validation runner
- Performance timing and reporting
- Clear success/failure messaging

### `/scripts/show-checks.js` (New)  
- Interactive script discovery
- Performance categorization
- Usage recommendations

## Testing
- [x] `typecheck:quick` works (~2s, modified files only)
- [x] `show-checks` displays all available scripts correctly
- [x] `check dev` runs quick validation (typecheck:quick + lint)
- [x] All new scripts are executable and functional

## Impact on Claude Code
Now Claude Code can:
1. Check available scripts with `npm run show-checks`
2. Use `npm run check:dev` for quick development validation  
3. Use `npm run check:pr` for PR review validation
4. Use `npm run check:full` only for comprehensive pre-deployment checks

This eliminates the timeout issues and dramatically improves development workflow speed.

🤖 Generated with [Claude Code](https://claude.ai/code)